tests: outbox: reorganize
